Move components for result verdict and flags into separate modules
