+++ assets/components/pages/project/assignments/AssignmentListItem.js
@@ -3,44 +3,10 @@ import React from "react";
 import { Link } from "react-router-dom";
 import { Item, List } from "semantic-ui-react";
 
-import verdicts, { verdictColors, verdictSymbols } from "../../../../constants/verdicts";
 import { CurationAssignmentPropType, CurationResultPropType } from "../../../propTypes";
+import Flags from "../../../result/Flags";
+import Verdict from "../../../result/Verdict";
 import VariantId from "../../../VariantId";
-
-const Verdict = ({ verdict }) => (
-  <span
-    style={{
-      color: verdictColors[verdict],
-    }}
-  >
-    {verdictSymbols[verdict]}
-  </span>
-);
-
-Verdict.propTypes = {
-  verdict: PropTypes.oneOf(verdicts).isRequired,
-};
-
-const Flags = ({ result }) => (
-  <span style={{ fontFamily: "monospace" }}>
-    {result.flag_mapping_error ? "M" : "_"}
-    {result.flag_genotyping_error ? "G" : "_"}
-    {result.flag_homopolymer ? "H" : "_"}
-    {result.flag_no_read_data ? "N" : "_"}
-    {result.flag_reference_error ? "E" : "_"}
-    {result.flag_strand_bias ? "S" : "_"}
-    {result.flag_mnp ? "P" : "_"}
-    {result.flag_essential_splice_rescue ? "R" : "_"}
-    {result.flag_minority_of_transcripts ? "T" : "_"}
-    {result.flag_weak_exon_conservation ? "C" : "_"}
-    {result.flag_last_exon ? "L" : "_"}
-    {result.flag_other_transcript_error ? "O" : "_"}
-  </span>
-);
-
-Flags.propTypes = {
-  result: CurationResultPropType.isRequired,
-};
 
 const CurationResult = ({ result }) => (
   <React.Fragment>
